+++ __tests__/controllers/search.test.js
@@ -1,6 +1,6 @@
 // module
-const { getAggregationQuery, getFilterExp,
-    getDateRangeArray, getDateRangeExp, getSortExp } = require('../../controllers/middleware/search')._test;
+const { getAggregationQuery, getDateRangeArray,
+    exp_matchFilter, exp_postFiledMask, exp_sortRule, exp_dateRange } = require('../../controllers/middleware/fetch')._test;
 
 
 
@@ -14,27 +14,18 @@ describe('Check the ENV', () => {
 
 describe('Bundle: search.js', () => {
     test('Fn: getAggregationQuery : Should construct Mongo query expression for search operations', () => {             // todo: throw to MongoDB for a test
-        expect(JSON.stringify(getAggregationQuery({ params: {}, session: {}, query: { date: {} }}))).toBe('[{\"$m' +
-            'atch\":{\"status\":{\"$eq\":\"published\"},\"visibility.hidden\":false}},{\"$sort\":{\"visibility.pi' +
-            'nned\":-1,\"time.updated\":-1}},{\"$group\":{\"_id\":null,\"count\":{\"$sum\":1},\"post\":{\"$push\"' +
-            ':\"$$ROOT\"}}},{\"$project\":{\"_id\":0,\"post\":{\"$slice\":[\"$post\",{\"$multiply\":[{\"$add\":[{' +
-            '\"$cond\":{\"if\":{\"$lt\":[1,{\"$ceil\":{\"$divide\":[\"$count\",null]}}]},\"then\":{\"$literal\":1' +
-            '},\"else\":{\"$ceil\":{\"$divide\":[\"$count\",null]}}}},-1]},null]},null]},\"meta\":{\"count\":\"$c' +
-            'ount\",\"num\":{},\"now\":{\"$cond\":{\"if\":{\"$lt\":[1,{\"$ceil\":{\"$divide\":[\"$count\",null]}}' +
-            ']},\"then\":{\"$literal\":1},\"else\":{\"$ceil\":{\"$divide\":[\"$count\",null]}}}},\"end\":{\"$ceil' +
-            '\":{\"$divide\":[\"$count\",null]}},\"baseUrl\":{},\"sort\":{\"$literal\":{\"visibility.pinned\":-1,' +
-            '\"time.updated\":-1}},\"date\":{\"$literal\":{}}}}}]');
-    });
-
-    test('Fn: getFilterExp: Should construct Mongo query expression (for $match)', () => {
-        expect(getFilterExp({ session: {}, params: {}, query: {} }))
-            .toEqual({ status: { $eq: 'published'}, 'visibility.hidden': false });
-        expect(getFilterExp({ session: {}, params: { search: {} }, query: {}}))
-            .toEqual({ $text: { $search: {} }, status: { $eq: 'published'}, 'visibility.hidden': false });
-        expect(getFilterExp({ session: {}, params: { category: 'test'}, query: {} }))
-            .toEqual({ category: 'test', status: { $eq: 'published'}, 'visibility.hidden': false });
-        expect(getFilterExp({ session: {}, params: {}, query: { date: '20180510-20190101/' } })['time.updated'])
-            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
+        expect(JSON.stringify(getAggregationQuery({ params: {}}, 0, null, null))).toBe('[{\"$match\":' +
+            '{\"status\":{\"$eq\":\"published\"},\"visibility.hidden\":false}},{\"$project\":{\"conte' +
+            'nt\":0}},{\"$sort\":{\"visibility.pinned\":-1,\"time.updated\":-1}},{\"$group\":{\"_id\"' +
+            ':null,\"count\":{\"$sum\":1},\"post\":{\"$push\":\"$$ROOT\"}}},{\"$project\":{\"_id\":0,' +
+            '\"post\":{\"$slice\":[\"$post\",{\"$multiply\":[{\"$add\":[{\"$cond\":{\"if\":{\"$lt\":[' +
+            '1,{\"$ceil\":{\"$divide\":[\"$count\",10]}}]},\"then\":{\"$literal\":1},\"else\":{\"$cei' +
+            'l\":{\"$divide\":[\"$count\",10]}}}},-1]},10]},10]},\"meta\":{\"count\":\"$count\",\"num' +
+            '\":{\"$literal\":10},\"now\":{\"$cond\":{\"if\":{\"$lt\":[1,{\"$ceil\":{\"$divide\":[\"$' +
+            'count\",10]}}]},\"then\":{\"$literal\":1},\"else\":{\"$ceil\":{\"$divide\":[\"$count\",1' +
+            '0]}}}},\"end\":{\"$ceil\":{\"$divide\":[\"$count\",10]}},\"sort\":{\"$literal\":{\"visib' +
+            'ility.pinned\":-1,\"time.updated\":-1}},\"route\":{\"$literal\":null},\"period\":{\"$lit' +
+            'eral\":{}}}}}]');
     });
 
     test('Fn: getDateRangeArray : Should translate into an array that contains a time range from a string', () => {
@@ -60,28 +51,46 @@ describe('Bundle: search.js', () => {
         expect(getDateRangeArray('20190101-20180510')   ).toEqual([[2019, 1, 1  ], [2018, 5, 10 ]]);
     });
 
-    test('Fn: getDateRangeExp : Should construct Mongo query expression based on a time range from an array', () => {
-        expect(getDateRangeExp([2018, 0, 0  ], [0   , 0, 0  ]))
-            .toEqual({ '$gte': new Date('2018-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-01T00:00:00.000Z') });
-        expect(getDateRangeExp([2018, 5, 0  ], [0   , 0, 0  ]))
-            .toEqual({ '$gte': new Date('2018-05-01T00:00:00.000Z'), '$lt': new Date('2018-06-01T00:00:00.000Z') });
-        expect(getDateRangeExp([2018, 5, 10 ], [0   , 0, 0  ]))
-            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2018-05-11T00:00:00.000Z') });
-        expect(getDateRangeExp([0   , 0, 0  ], [2019, 1, 1  ]))
-            .toEqual({ '$gte': new Date('2019-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
-        expect(getDateRangeExp([2018, 0, 0  ], [2019, 1, 1  ]))
-            .toEqual({ '$gte': new Date('2018-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
-        expect(getDateRangeExp([2018, 5, 0  ], [2019, 1, 1  ]))
-            .toEqual({ '$gte': new Date('2018-05-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
-        expect(getDateRangeExp([2018, 5, 10 ], [2019, 1, 1  ]))
-            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
-        expect(getDateRangeExp([2019, 1, 1  ], [2018, 5, 10 ]))
+    test('Fn: exp_matchFilter: Should construct Mongo query expression (for $match)', () => {
+        expect(exp_matchFilter({}, {}))
+            .toEqual({ status: { $eq: 'published'}, 'visibility.hidden': false });
+        expect(exp_matchFilter({ search: {} }, {}))
+            .toEqual({ $text: { $search: {} }, status: { $eq: 'published'}, 'visibility.hidden': false });
+        expect(exp_matchFilter({ category: 'test'}, {}))
+            .toEqual({ category: 'test', status: { $eq: 'published'}, 'visibility.hidden': false });
+        expect(exp_matchFilter({}, { date: '20180510-20190101/' })['time.updated'])
             .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
     });
 
-    test('Fn: getSortExp : Should construct Mongo query expression (for $sort)', () => {
-        expect(getSortExp({ 'time.updated': 0 }, {})).toEqual({ 'time.updated': -1, 'visibility.pinned': -1 });
-        expect(getSortExp({ 'time.updated': 1 }, {})).toEqual({ 'time.updated': 1, 'visibility.pinned': -1 });
-        expect(getSortExp({}, {})).toEqual({ 'time.updated': -1, 'visibility.pinned': -1 });
+    test('Fn: exp_postFiledMask - Should construct Mongo query expression (for $project)', () => {
+        expect(exp_postFiledMask({ stackType: 'posts' }))
+            .toEqual({ _id: 1, canonical: 1, title: 1, author: 1, category: 1, tags: 1, time: 1 });
+        expect(exp_postFiledMask({}))
+            .toEqual({ content: 0 });
+    });
+
+    test('Fn: exp_sortRule - Should construct Mongo query expression (for $sort)', () => {
+        expect(exp_sortRule({ 'time.updated': 0 })).toEqual({ 'time.updated': -1, 'visibility.pinned': -1 });
+        expect(exp_sortRule({ 'time.updated': 1 })).toEqual({ 'time.updated': 1, 'visibility.pinned': -1 });
+        expect(exp_sortRule({})).toEqual({ 'time.updated': -1, 'visibility.pinned': -1 });
+    });
+
+    test('Fn: exp_dateRange - Should construct Mongo query expression based on a time range from an array', () => {
+        expect(exp_dateRange([2018, 0, 0  ], [0   , 0, 0  ]))
+            .toEqual({ '$gte': new Date('2018-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-01T00:00:00.000Z') });
+        expect(exp_dateRange([2018, 5, 0  ], [0   , 0, 0  ]))
+            .toEqual({ '$gte': new Date('2018-05-01T00:00:00.000Z'), '$lt': new Date('2018-06-01T00:00:00.000Z') });
+        expect(exp_dateRange([2018, 5, 10 ], [0   , 0, 0  ]))
+            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2018-05-11T00:00:00.000Z') });
+        expect(exp_dateRange([0   , 0, 0  ], [2019, 1, 1  ]))
+            .toEqual({ '$gte': new Date('2019-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
+        expect(exp_dateRange([2018, 0, 0  ], [2019, 1, 1  ]))
+            .toEqual({ '$gte': new Date('2018-01-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
+        expect(exp_dateRange([2018, 5, 0  ], [2019, 1, 1  ]))
+            .toEqual({ '$gte': new Date('2018-05-01T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
+        expect(exp_dateRange([2018, 5, 10 ], [2019, 1, 1  ]))
+            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
+        expect(exp_dateRange([2019, 1, 1  ], [2018, 5, 10 ]))
+            .toEqual({ '$gte': new Date('2018-05-10T00:00:00.000Z'), '$lt': new Date('2019-01-02T00:00:00.000Z') });
     });
 });