+++ routers/page.js
@@ -1,5 +1,5 @@
 const { Device } = require('../controllers/engines/router');
-const search = require('../controllers/middleware/search');
+const { fetch } = require('../controllers/middleware/fetch');
 
 
 
@@ -11,7 +11,7 @@ const PageRouter = new Device([{
     controller:     (req, res) => res.render('./theme'),
 }, {
     route:          '/search/:search',
-    controller:     search.find(),
+    controller:     (req, res, next) => fetch({ num: res.locals._site.sets.num })(req, res, next),
     settings:       { crawler: false, template: './theme/post/search' },
 }]);